+++ packages/expo-media-library/android/src/main/java/expo/modules/medialibrary/next/permissions/enums/AccessPrivileges.kt
@@ -1,0 +1,7 @@
+package expo.modules.medialibrary.next.permissions.enums
+
+enum class AccessPrivileges(val value: String) {
+  ALL("all"),
+  LIMITED("limited"),
+  NONE("none")
+}